server/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
beanie==1.29.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dnspython==1.16.0
fastapi==0.115.8
h11==0.14.0
idna==3.10
lazy-model==0.2.0
motor==3.7.0
passlib==1.7.4
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.45.3
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.34.0
